Added MultiProcessing, Modified parents generation
Added multiprocessing pools for calling the java program. Set pool_size
to the number of logical cpu's on your laptop/desktop.

Changed the way parents are generated from the initial seed, with new
weights being generated from a delta of a seed weight of 0.1. Mutations
are also handled this way.